Split web workspace session tests
